requirements.txt
asgiref==3.6.0
black==23.1.0
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.7
django-cors-headers==3.14.0
django-environ==0.10.0
django-excel==0.0.10
djangorestframework==3.14.0
docker==6.0.1
drf-yasg==1.21.5
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
lml==0.1.0
MarkupSafe==2.1.2
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.1
Pillow==9.4.0
platformdirs==3.1.1
psycopg2-binary==2.9.5
pyexcel==0.7.0
pyexcel-io==0.6.6
pyexcel-webio==0.1.4
pytz==2022.7.1
requests==2.28.2
ruamel.yaml==0.17.21
sqlparse==0.4.3
texttable==1.6.7
uritemplate==4.1.1
urllib3==1.26.15
websocket-client==1.5.1
